% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demo Date}
\format{
## `demo_data`
A list of matrices and data frames
}
\usage{
demo_data
}
\description{
A collection of data structures for experimenting with {tilemakr} functions.
}
\keyword{datasets}
