% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_circle.R
\name{gen_circle}
\alias{gen_circle}
\title{Generate Circle}
\usage{
gen_circle(group_id, size, padding, start)
}
\arguments{
\item{group_id}{A unique identifier for the series of polygon data points
creating a tile shape. Derived from the matrix layout.}

\item{size}{An integer detailing the radius of the circle tile shape.}

\item{padding}{An integer detailing the space between surrounding circle
tile shapes.}

\item{start}{A vector of length two comprised of the x,y position for the next
hexagon tile shape to draw.}
}
\value{
A two element list with the shape data points and coordinates for the
next circle tile shapes to draw.
}
\description{
`gen_circle()` creates a series of polygon data points for an individual
hexagon tile group.
}
\examples{
gen_circle("A", 5, 2, c(0,0))
}
\seealso{
[make_circle_tiles(), make_tiles()]
}
