% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_diamond.R
\name{gen_diamond}
\alias{gen_diamond}
\title{Generate Diamond}
\usage{
gen_diamond(group_id, size, padding, start)
}
\arguments{
\item{group_id}{A unique identifier for the series of polygon data points
creating a tile shape. Derived from the matrix layout.}

\item{size}{An integer detailing the point radius of the diamond tile shape.}

\item{padding}{An integer detailing the space between surrounding diamond
tile shapes.}

\item{start}{A vector of length two comprised of the x,y position for the next
diamond tile shape to draw.}
}
\value{
A two element list with the shape data points and coordinates for the
next diamond tile shapes to draw.
}
\description{
`gen_diamond()` creates a series of polygon data points for an individual
diamond tile group.
}
\examples{
gen_diamond("A", 5, 2, c(0,0))
}
\seealso{
[make_diamond_tiles(), make_tiles()]
}
