% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_square.R
\name{gen_square}
\alias{gen_square}
\title{Generate Square}
\usage{
gen_square(group_id, size, padding, start)
}
\arguments{
\item{group_id}{A unique identifier for the series of polygon data points
creating a tile shape. Derived from the matrix layout.}

\item{size}{An integer detailing the length and width of the square tile shape.}

\item{padding}{An integer detailing the space between surrounding square
tile shapes.}

\item{start}{A vector of length two comprised of the x,y position for the next
square tile shape to draw.}
}
\value{
A two element list with the shape data points and coordinates for the
next square tile shapes to draw.
}
\description{
`gen_square()` creates a series of polygon data points for an individual
tile group.
}
\examples{
gen_square("A", 10, 2, c(0,0))
}
\seealso{
[make_square_tiles(), make_tiles()]
}
