% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_tile_data}
\alias{geo_tile_data}
\title{List of matrices to create tile maps}
\format{
## `geo_tile_data`
A list of matrices for creating tile maps using any shape supported
by the {tilemaker} functions.
}
\usage{
geo_tile_data
}
\description{
Pre-generated layout matrices for quick tile making, then join to and plot
with your own data sets.
}
\keyword{datasets}
