% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_diamond.R
\name{make_diamond_tiles}
\alias{make_diamond_tiles}
\title{Make Diamond Tile Data}
\usage{
make_diamond_tiles(layout, size, padding)
}
\arguments{
\item{layout}{A matrix outlining the desired tile map layout. Each position
intended for a tile should have a unique identifier while all empty positions
should feature a '0'.}

\item{size}{An integer detailing the point radius of the diamond tile shape.}

\item{padding}{An integer detailing the space between surrounding diamond
tile shapes.}
}
\value{
A data frame of coordinates for drawing the diamond tiles and center
coordinates for plotting the group text.
}
\description{
`make_diamond_tiles()` creates a data frame of tile data points for each
valid element observed in the layout.
}
\examples{
layout = matrix(
  data = c(
    "A", 0, 0, "B",
    0, "C", "D", "E",
    0, 0, "F", 0,
    0, "G", 0, 0
  ),
  byrow = TRUE,
  nrow = 4,
  ncol = 4
)
make_diamond_tiles(layout, 10, 2)
}
\seealso{
[gen_diamond(), make_tiles()]
}
