% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hex.R
\name{make_hex_tiles}
\alias{make_hex_tiles}
\title{Make Hexagon Tile Data}
\usage{
make_hex_tiles(layout, size, padding)
}
\arguments{
\item{layout}{A matrix outlining the desired tile map layout. Each position
intended for a tile should have a unique identifier while all empty positions
should feature a '0'.}

\item{size}{An integer detailing the point radius of the hexagon tile shape.}

\item{padding}{An integer detailing the space between surrounding hexagon
tile shapes.}
}
\value{
A data frame of coordinates for drawing the hexagon tiles and center
coordinates for plotting the group text.
}
\description{
`make_hex_tiles()` creates a data frame of tile data points for each
valid element observed in the layout.
}
\examples{
layout <- demo_data[["Example Matrix"]]

make_hex_tiles(layout, 10, 2)
}
\seealso{
[gen_hex(), make_tiles()]
}
