% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{make_tiles}
\alias{make_tiles}
\title{Make Tiles}
\usage{
make_tiles(layout, type = c("square", "hexagon"), size = 10, padding = 2)
}
\arguments{
\item{layout}{A matrix outlining the desired tile map layout. Each position
intended for a tile should have a unique identifier while all empty positions
should feature a '0'.}

\item{type}{A string identifying the shape type to generate; either a 'square'
or 'hexagon'.}

\item{size}{An integer detailing the length and width of the square tile shape.}

\item{padding}{An integer detailing the space between surrounding square
tile shapes.}
}
\value{
A data frame of coordinates to draw tile maps.
}
\description{
`make_tiles()` creates data frame of coordinates to draw square or hexagon
tile maps which can be plotted in {ggplot2}.
}
\examples{
layout = matrix(
  data = c(
    "A", 0, 0, "B",
    0, "C", "D", "E",
    0, 0, "F", 0,
    0, "G", 0, 0
  ),
  byrow = TRUE,
  nrow = 4,
  ncol = 4
)
make_tiles(layout, 'square', 10, 2)
}
\seealso{
[make_square_tiles(), gen_square()]
}
