% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{plot_tiles}
\alias{plot_tiles}
\title{Plot Tiles}
\usage{
plot_tiles(df, labels = TRUE)
}
\arguments{
\item{df}{A data frame of coordinates}

\item{labels}{A boolean (logical) value indicating if labels should be plotted.}
}
\value{
A {ggplot2} object
}
\description{
`plot_tiles()` takes a data frame of coordinates and to draw the corresponding
shapes to a {ggplot2} object.
}
\examples{
# Prepare coordinates dataframe
df_matrix = geo_tile_data[['US States']]
df_coord = make_tiles(df_matrix, "square")

# Show the tile map
plot_tiles(df_coord)

# Without labels
plot_tiles(df_coord, fALSE)
}
