% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{preview_shapes}
\alias{preview_shapes}
\title{Preview Shapes}
\usage{
preview_shapes()
}
\value{
A `ggplot2` object
}
\description{
`preview_shapes()` shows a preview of what tile maps would look like with
different shapes.
}
\examples{
preview_shapes()
}
