% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sales_data}
\alias{sales_data}
\title{Tabular layout for tile map}
\format{
## `sales_data`
A data frame with 7 rows and 2 columns:
\describe{
  \item{id}{Unique identifiers for the tile shapes}
  \item{sales}{Sales figures for each tile shape}
}
}
\usage{
sales_data
}
\description{
A data set intended to illustrate how to join generated tile map data to
a corresponding data set with variables of interest.
}
\keyword{datasets}
