% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabular_layout}
\alias{tabular_layout}
\title{Tabular Layout for Tile Map}
\format{
## `tabular_layout`
A data frame with 4 rows and 4 columns:
\describe{
  \item{V1}{Layout elements}
  \item{V2}{Layout elements}
  \item{V3}{Layout elements}
  \item{V4}{Layout elements}
}
}
\usage{
tabular_layout
}
\description{
A grid style layout for creating a tile map. Each column and row describes
elements to great the map. This data type should be converted to a matrix
using `tabular_to_matrix()`.
}
\keyword{datasets}
