% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tabular_to_matrix}
\alias{tabular_to_matrix}
\title{Tabular to Matrix}
\usage{
tabular_to_matrix(df)
}
\arguments{
\item{df}{A data frame where rows and columns represent a grid layout for
the tile map creation. Column names/headers should be excluded.}
}
\value{
A matrix form of the layout intended for a tile maps.
}
\description{
`tabular_to_matrix()` converts a layout stored in a data frame to that of a
matrix for proper use in the `make_tiles()` function. Remember to exclude
column names (or remove the header) from the data frame.
}
\examples{
layout <- demo_data[["Example Data Frame"]]

# Currently a data frame
class(layout)

new_layout = tabular_to_matrix(layout)

# Now a matrix
class(new_layout)
}
\seealso{
[make_tiles()]
}
