% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tile_layouts}
\alias{tile_layouts}
\title{Tile Layouts}
\format{
A list of matrices, which are:
\describe{
\item{`Example Matrix`}{A matrix featuring an example, desired layout}
\item{`Example Data Frame`}{A data frame featuring the desired layout in a data frame object}
\item{`Sales Data`}{A data frame featuring sales data associated with shapes id's from `Example Matrix` and `Example Data Frame`}
\item{`NBA Standings`}{A data frame featuring win/loss records associated with shape id's from `NBA Franchises`}
}
}
\usage{
tile_layouts
}
\description{
Pre-generated layout matrices for quick tile making, then join to and plot
with your own data sets.
}
\keyword{datasets}
