% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{make_tiles}
\alias{make_tiles}
\title{Make Tiles}
\usage{
make_tiles(
  layout,
  type = c("square", "hexagon", "circle", "diamond"),
  size = 10,
  padding = 2
)
}
\arguments{
\item{layout}{A matrix outlining the desired tile map layout. Each position
intended for a tile should have a unique identifier while all empty positions
should feature a '0'.}

\item{type}{A string identifying the shape type to generate; 'square', 'hexagon',
'circle', or 'diamond'.}

\item{size}{An integer detailing the length and width of the square tile shape.}

\item{padding}{An integer detailing the space between surrounding tile shapes.}
}
\value{
A data frame of coordinates to draw tile maps.
}
\description{
`make_tiles()` creates data frame of coordinates to draw square or hexagon
tile maps which can be plotted in {ggplot2}.
}
\examples{
layout = demo_data[["Example Matrix"]]

# Square tile map
make_tiles(layout, 'square', 10, 2)

# Hexagon tile map
make_tiles(layout, 'hexagon')

# Circle tile map
make_tiles(layout, 'circle')
}
\seealso{
[make_square_tiles(), make_hex_tiles(), make_circle_tiles(), make_diamond_tiles()]
}
