% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tile_layouts}
\alias{tile_layouts}
\title{List of matrices to create tile maps}
\format{
## `tile_layouts`
A list of matrices for creating tile maps using any shape supported
by the {tilemaker} functions.
}
\usage{
tile_layouts
}
\description{
Pre-generated layout matrices for quick tile making, then join to and plot
with your own data sets.
}
\keyword{datasets}
